requirements.txt
aiohttp==1.0.5
async-timeout==3.0.0
atomicwrites==1.1.5
attrs==18.1.0
certifi==2018.4.16
chardet==3.0.4
discord.py==0.16.12
Django==2.1
httmock==1.2.6
idna==2.7
more-itertools==4.3.0
multidict==4.3.1
pluggy==0.7.1
py==1.5.4
pytest==3.7.1
pytest-asyncio==0.9.0
pytest-django==3.3.3
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
six==1.11.0
urllib3==1.23
websockets==3.4
